isGenericMethod() 위치 변경
MethodCallInfo -> AbstractMethodInfo
